update home and modles
