+++ src/components/Home.tsx
@@ -1,65 +1,26 @@
-import { useEffect, useState } from 'react';
+import { useContext } from 'react';
+import { HistoryBarContext, HistoryBarContextType } from './SettingsBarContext';
 
 import WrapperFavorites from './WrapperFavorites';
-import SmallInfoCard from './SmallInfoCard';
+import WrapperLastSeen from './WrapperLastSeen';
+import WrapperPopular from './WrapperPopular';
 import MovieBanner from './MovieBanner';
-import Container from './Container';
-import Slider from './Slider';
-import Title from './Title';
-
-import api from 'src/service/api';
-
-export type MoviesType = {
-  vote_count: number;
-  vote_average: number;
-  original_title: string;
-  poster_path: string;
-  overview: string;
-  title: string;
-  id: number;
-}
 
 const Home = () => {
-  const [movies, setMovies] = useState<[]>([]);
-  const [history, setHistory] = useState<MoviesType[]>();
-
-  useEffect(() => {
-    api.getPopularQueryList().then(res => setMovies(res.slice(0, 10)));
-    api.getDataByIds(api.getHistoryIdList()).then((res) => { setHistory(res) });
-  }, []);
+  const { historyState, favoriteState }: HistoryBarContextType = useContext(HistoryBarContext);
 
   return (
     <>
       <MovieBanner />
       <div className="wrapper-space">
-        <WrapperFavorites />
+        {favoriteState ? <WrapperFavorites /> : null}
       </div>
-      <Container>
-        {history && <div className="wrapper-space">
-          {history.length ? <><Title text={"Last seens"} />
-            <Slider>
-              {history.map((movie: MoviesType) => {
-                return (
-                  <div className="slide" key={movie.poster_path}>
-                    <SmallInfoCard id={movie.id} />
-                  </div>
-                );
-              })}
-            </Slider> </> : null}
-        </div>}
-        <div className="wrapper-space">
-          <Title text={'Popular movies'} />
-          {movies.length ? <Slider>
-            {movies.map((movie: MoviesType) => {
-              return (
-                <div className="slide" key={movie.poster_path}>
-                  <SmallInfoCard id={movie.id} />
-                </div>
-              );
-            })}
-          </Slider> : null}
-        </div>
-      </Container>
+      <div className="wrapper-space">
+        {historyState ? <WrapperLastSeen /> : null}
+      </div>
+      <div className="wrapper-space">
+        <WrapperPopular />
+      </div>
     </>
   );
 };
